pyproject.toml
[tool.poetry]
name = "moobot"
version = "0.0.1"
description = "Private server management bot."
authors = ["Stephan Lensky <public@slensky.com>"]
readme = "README.md"
license = "AGPL-3.0-only"
homepage = "https://github.com/stephanlensky/moobot"
repository = "https://github.com/stephanlensky/moobot"
include = [
    "LICENSE",
]
packages = [
    {include = "moobot"}
]

[tool.poetry.scripts]
moobot = "moobot._entry:run_discord_bot"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
requests = "^2.27.1"
"discord.py" = { git = "https://github.com/Rapptz/discord.py.git", rev = "277d35c" }
python-dateutil = "^2.8.2"
pydantic = {extras = ["dotenv"], version = "^1.9.0"}
SQLAlchemy = "^1.4.35"
wrapt = "^1.14.0"
psycopg2 = "^2.9.3"
sqlmodel = "^0.0.6"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pytest-mock = "^3.7.0"
pytest-asyncio = "^0.18.3"
black = "^22.3.0"
isort = "^5.10.1"
mypy = "^0.961"
pylint = "^2.13.3"
docformatter = "^1.4"
types-requests = "^2.27.15"
SQLAlchemy = {extras = ["mypy"], version = "^1.4.35"}
boto3-stubs = {extras = ["s3"], version = "^1.23.8"}
mkdocs-material = "^8.3.6"
types-PyYAML = "^6.0.10"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.master]
disable="""
C0103,C0114,C0115,C0116,C0200,C0330,C0326,W0703,W0105,R1705,R0901,R0902,R0903,R0904,R0911,R0912,R0913,R0914,
R0915,R0916,R1702,C0206,R0801,R0201,W1203,W0706,R0401,C0104,C0123,C0301"""
extension-pkg-whitelist="pydantic"

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100

[tool.black]
line-length = 100

[tool.mypy]
show_column_numbers = true
disallow_untyped_defs = true
plugins = ["sqlalchemy.ext.mypy.plugin"]

[[tool.mypy.overrides]]
module = [
    "craigslist",
    "geopy.*",
    "boolean",
    "discord.*",
    "wrapt",
    "apscheduler.*",
    "geopandas.*",
    "pandas.*",
    "scipy.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
    # geopandas has not yet switched off of distutils
    'ignore:distutils Version classes are deprecated:DeprecationWarning'
]
